+++ src/math/interpolations/linear.js
@@ -1,0 +1,4 @@
+export const linear = points => points.reduce(
+	(acc, x) => `${acc}${acc ? 'L' : 'M'}${x[0]},${x[1]}`,
+	''
+);