+++ src/plotery.js
@@ -3,6 +3,7 @@ export { Surface } from './surface.component';
 export { CartesianAxis } from './cartesian/cartesian-axis.component';
 export { LinearAxis } from './cartesian/linear-axis.component';
 export { LogAxis } from './cartesian/log-axis.component';
+export { CardinalLine } from './cartesian/cardinal-line.component';
 export { LinearLine } from './cartesian/linear-line.component';
 export { BarLine } from './cartesian/bar-line.component';
 export { Zoom } from './interactions/zoom.component';